# docs/architecture/control-plane/operator-ui-mvp-openapi.yaml
openapi: 3.1.0
info:
  title: OneFlow Control Plane - Monitoring First API
  version: 0.1.0
  description: |
    Current contract for the monitoring-first control-plane starter implemented by
    com.etl.controlplane.ControlPlaneApiApplication.
servers:
  - url: /api/v1
    description: Relative base path for the monitoring-first control-plane API
...
paths:
...
  /jobs/{jobKey}:
    get:
      tags: [Jobs]
      operationId: getJobByKey
      summary: Get one aggregated job detail payload
      parameters:
        - $ref: '#/components/parameters/jobKey'
      responses:
        '200':
          description: Aggregated job detail with recent runs and trigger events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobBundleDetailResponse'
        '404':
          description: Job bundle not found
  /jobs/{jobKey}/config:
    get:
      tags: [Jobs]
      operationId: getJobConfigByKey
      summary: Get one read-only job-config.yaml payload
      parameters:
        - $ref: '#/components/parameters/jobKey'
      responses:
        '200':
          description: Read-only job-config payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobBundleConfigResponse'
        '404':
          description: Job bundle not found
  /jobs/{jobKey}:trigger-now:
    post:
      tags: [Jobs]
      operationId: triggerJobNow
      summary: Record an accepted placeholder trigger decision for a known job bundle
      parameters:
        - $ref: '#/components/parameters/jobKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerNowRequest'
      responses:
        '202':
          description: Placeholder trigger decision accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerNowDecisionResponse'
        '404':
          description: Job bundle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerNowDecisionResponse'
  /jobs/{jobKey}/trigger-events:
    get:
      tags: [Jobs]
      operationId: listJobTriggerEvents
      summary: List recent trigger events for one job bundle
      parameters:
        - $ref: '#/components/parameters/jobKey'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
          description: Optional alias that takes precedence over `limit`.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Zero-based page number.
      responses:
        '200':
          description: Trigger event history entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerEventListResponse'
        '404':
          description: Job bundle not found
...
components:
  parameters:
    jobKey:
      in: path
      name: jobKey
      required: true
      schema:
        type: string
...
  schemas:
    ReadinessStatus:
      type: string
      enum: [READY, INACTIVE, INVALID]
    JobBundleSummaryView:
      type: object
      properties:
        jobKey:
          type: string
        displayName:
          type: string
        jobConfigPath:
          type: string
        readinessStatus:
          $ref: '#/components/schemas/ReadinessStatus'
        validationMessages:
          type: array
          items:
            type: string
      required: [jobKey, displayName, jobConfigPath, readinessStatus, validationMessages]
...
    JobBundleDetailResponse:
      type: object
      properties:
        job:
          $ref: '#/components/schemas/JobBundleSummaryView'
        recentRuns:
          type: array
          items:
            $ref: '#/components/schemas/RunSummaryView'
        recentTriggerEvents:
          type: array
          items:
            $ref: '#/components/schemas/TriggerEventView'
      required: [job, recentRuns, recentTriggerEvents]
    JobBundleConfigResponse:
      type: object
      properties:
        jobKey:
          type: string
        displayName:
          type: string
        jobConfigPath:
          type: string
        rawYaml:
          type: string
      required: [jobKey, displayName, jobConfigPath, rawYaml]
    RunSummaryView:
      type: object
      properties:
        scenario:
          type: string
        jobExecutionId:
          oneOf:
            - type: integer
              format: int64
            - type: 'null'
        runMode:
          oneOf:
            - type: string
            - type: 'null'
        recoveryPolicy:
          oneOf:
            - type: string
            - type: 'null'
        status:
          type: string
        startTime:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        endTime:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        durationSeconds:
          oneOf:
            - type: integer
              format: int64
            - type: 'null'
        sourceCount:
          oneOf:
            - type: integer
              format: int64
            - type: 'null'
        writtenCount:
          oneOf:
            - type: integer
              format: int64
            - type: 'null'
        rejectedCount:
          oneOf:
            - type: integer
              format: int64
            - type: 'null'
        logPath:
          type: string
      required: [scenario, jobExecutionId, runMode, recoveryPolicy, status, startTime, endTime, durationSeconds, sourceCount, writtenCount, rejectedCount, logPath]
...
    TriggerNowRequest:
      type: object
      properties:
        reason:
          type: string
        requestedBy:
          type: string
    TriggerNowDecisionStatus:
      type: string
      enum: [ACCEPTED, NOT_FOUND]
    TriggerNowDecisionResponse:
      type: object
      properties:
        jobKey:
          type: string
        decisionStatus:
          $ref: '#/components/schemas/TriggerNowDecisionStatus'
        message:
          type: string
        triggerEventId:
          oneOf:
            - type: string
            - type: 'null'
      required: [jobKey, decisionStatus, message, triggerEventId]
    TriggerEventView:
      type: object
      properties:
        triggerEventId:
          type: string
        jobKey:
          type: string
        decisionStatus:
          type: string
        reason:
          type: string
        requestedBy:
          type: string
        requestedAt:
          type: string
          format: date-time
        launchedRunId:
          oneOf:
            - type: string
            - type: 'null'
        message:
          type: string
      required: [triggerEventId, jobKey, decisionStatus, reason, requestedBy, requestedAt, launchedRunId, message]
    TriggerEventListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TriggerEventView'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
      required: [items, page, size, totalItems]
...